Refactor obliterate tests out of conflictFarm
